Feature: /track now accepts inactivity gaps as optional argument
The inactivity gap (i.e. how long since the last station was heard)
was set to a fixed ammount of hours (1h, now 6h). This is an important
heuristic: the user does not want to get a message for every new APRS
packet heard from the ISS, but rather only when it has not seen
activity for a while. Now, the user can input a custom gap in seconds.